+++ packages/core/legos/src/sources.ts
@@ -80,7 +80,6 @@ import * as ncfx from '@chainlink/ncfx-adapter'
 import * as nikkei from '@chainlink/nikkei-adapter'
 import * as nomics from '@chainlink/nomics-adapter'
 import * as oilpriceapi from '@chainlink/oilpriceapi-adapter'
-import * as onchain from '@chainlink/onchain-adapter'
 import * as onchain_gas from '@chainlink/onchain-gas-adapter'
 import * as openexchangerates from '@chainlink/openexchangerates-adapter'
 import * as orchid_adapter from '@chainlink/orchid-bandwidth-adapter'
@@ -199,7 +198,6 @@ export default {
   nikkei,
   nomics,
   oilpriceapi,
-  onchain,
   onchain_gas,
   openexchangerates,
   orchid_adapter,