+++ packages/sources/onchain/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export * as price from './price'